requirements.txt
Pillow==6.2.1
dhash==1.3
piexif==1.1.3
pybktree==1.1
python-xmp-toolkit==2.0.1
